refactor(react/Form): move/extract some form elements
