+++ packages/react/src/components/Form/TextField.tsx
@@ -1,13 +1,10 @@
 import '@compassion-gds/css/src/components/Form/common.css';
 
-import Input, { InputProps } from './Input';
+import { FieldContainer, Input, InputProps } from './FormElements';
 
-interface TextFieldContainerProps extends InputProps {
+export interface TextFieldProps extends InputProps {
   label?: string;
   hint?: string,
-}
-
-export interface TextFieldProps extends TextFieldContainerProps {
   fieldClassName?: string,
 }
 
@@ -21,30 +18,9 @@ const TextField = ({
   ...rest
 }: TextFieldProps) => {
   return (
-    <TextFieldContainer id={id} label={label} state={state} hint={hint} className={className}>
+    <FieldContainer id={id} label={label} state={state} hint={hint} className={className}>
       <Input id={id} state={state} className={fieldClassName} {...rest} />
-    </TextFieldContainer>
-  );
-};
-
-export const TextFieldContainer = ({
-  id,
-  label = '',
-  state,
-  hint = '',
-  className,
-  children,
-}: TextFieldContainerProps) => {
-  const containerClassNames =  ['form-field-row', className].join(' ');
-  const labelClassNames = ['form-label', state].join(' ');
-  const helperClassNames = ['form-hint', state].join(' ');
-
-  return (
-    <div className={containerClassNames}>
-      {label && <label className={labelClassNames} htmlFor={id}>{label}</label>}
-      {children}
-      {hint && <span className={helperClassNames}>{hint}</span>}
-    </div>
+    </FieldContainer>
   );
 };
