+++ packages/react/src/components/Form/index.ts
@@ -1,2 +1,2 @@
-export { default as Input } from './Input';
+export { Input } from './FormElements';
 export { default as TextField } from './TextField';
